#14, Add checkbox to show/hide deprecated files
This adds a checkbox to the detailed view of a published files (its
version history) to toggle the visibility of deprecated files. This is
done through a user setting (`show_deprecated_files`) that can then be
checked by a filter_publishes hook to see if deprecated files need to be
filtered out.

Closes #1.